feat(search) Add search options for linked_ticket
Add linked_ticket as a search filter for both the django based search
and as a pre-filter for snuba based search.

A request we've heard from customers is the need to find which issues in
sentry have not been triaged into their external bug tracker. This new
filter makes it possible to find groups that have or don't have
integration based issues attached to them. Plugin based issue trackers
are not covered by these changes as the `GroupMeta` data is stored in
a way that makes searching it slow.

Refs APP-894